+++ src/common/base/runnable.ts
@@ -1,0 +1,3 @@
+/** A function that just runs; it takes no values and returns nothing. */
+export type Runnable = () => void;
+
